Move __mocks__ back to root
