+++ __mocks__/linter-provider-not-a-provider.ts
